requirements.txt
flask
openpyxl
python-barcode
pillow
gunicorn
